User Can Now Update Note
Now all of the CRUD Functions that Notes App needs To Have is Done.

+++ Notes Frontend/Notes Frontend/Home.swift
@@ -11,9 +11,14 @@ struct Home: View {
     @State var notes = [Note]()
     
     @State var showAdd = false
+    @State var isEditMode: EditMode = .inactive
     
     @State var showAlert = false
+    
     @State var deleteItem: Note?
+    @State var updateNote = ""
+    @State var updateNoteId = ""
+    
     
     var alert: Alert {
         Alert(title: Text("Delete"), message: Text("Are You Sure You Want To Delete This Note?"), primaryButton: .destructive(Text("Delete"), action: deleteNote), secondaryButton: .cancel())
@@ -22,28 +27,62 @@ struct Home: View {
     var body: some View {
         NavigationView {
             List(self.notes) { note in
-                Text(note.note)
-                    .padding(10)
-                    .onLongPressGesture {
-                        self.showAlert.toggle()
-                        deleteItem = note
+                if(self.isEditMode == .inactive) {
+                    Text(note.note)
+                        .padding()
+                        .onLongPressGesture {
+                            self.showAlert.toggle()
+                            deleteItem = note
+                        }
+                }else {
+                    HStack {
+                        Image(systemName: "pencil.circle.fill")
+                            .foregroundColor(.yellow)
+                        Text(note.note)
+                            .padding()
                     }
+                    .onTapGesture {
+                        self.updateNote = note.note
+                        self.updateNoteId = note.id
+                        self.showAdd.toggle()
+                    }
+                }
             }
             .alert(isPresented: $showAlert, content: {
                 alert
             })
             .sheet(isPresented: $showAdd, onDismiss: fetchNotes ,content: {
-                AddNoteView()
+                if (self.isEditMode == .inactive) {
+                    AddNoteView()
+                }else {
+                    UpdateNoteView(text: $updateNote, noteId: $updateNoteId)
+                }
             })
             .onAppear(perform: {
                     fetchNotes()
                 })
             .navigationTitle("Notes")
-            .navigationBarItems(trailing: Button(action: {
-                    self.showAdd.toggle()
-                }, label: {
-                    Text("Add")
-                }))
+            .navigationBarItems(
+                leading:
+                    Button(action: {
+                        if (self.isEditMode == .inactive) {
+                                self.isEditMode = .active
+                            }else {
+                                self.isEditMode = .inactive
+                            }
+                        }, label: {
+                            if(self.isEditMode == .inactive) {
+                                Text("Edit")
+                            }else {
+                                Text("Done")
+                            }
+                    }),
+                trailing:
+                    Button(action: {
+                            self.showAdd.toggle()
+                        }, label: {
+                            Text("Add")
+                        }))
         }
     }
     
@@ -62,6 +101,10 @@ struct Home: View {
         }
         
         task.resume()
+        
+        if (self.isEditMode == .active) {
+            self.isEditMode = .inactive
+        }
     }
     
     func deleteNote() {
